making UA allow to input decimal numbers

+++ public/js/app.js
@@ -278,7 +278,7 @@ export async function createNewProject() {
   errMsg('createError', '');
 
   const mode     = document.querySelector('input[name="createAssessMode"]:checked')?.value || 'point';
-  const sizeM    = parseInt($('createPlotSizeM').value) || 30;
+  const sizeM    = parseFloat($('createPlotSizeM').value) || 30;
   const boxM     = _parsePointBoxSize($('createPointBoxSizeM').value);
   const grid     = $('createSubGrid').value || '5x5';
   const aggRule  = $('createAggRule').value || 'majority';
@@ -369,7 +369,7 @@ function _parsePointBoxSize(v) {
 export async function saveProjectSettings() {
   if (!state.project) return;
   const mode    = $('settingsAssessMode').value;
-  const sizeM   = parseInt($('settingsPlotSizeM').value) || 30;
+  const sizeM   = parseFloat($('settingsPlotSizeM').value) || 30;
   const boxM    = _parsePointBoxSize($('settingsPointBoxSizeM').value);
   const grid    = $('settingsSubGrid').value || '5x5';
   const aggRule = $('settingsAggRule').value || 'majority';
